refactor: extra `useStableSearchWords`
